requirements.txt
numpy==1.18.5
pandas==1.0.5
Pillow==7.2.0
matplotlib==3.2.2
regex==2020.7.14
advertools==0.10.6
emoji==0.5.4
emojis==0.5.1
scikit-learn==0.23.1
scipy==1.4.1
Scrapy==2.3.0
seaborn==0.10.1
sklearn==0.0
gensim==3.8.3
torch==1.6.0
flair==0.6
joblib==0.16.0
tensorflow==2.3.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
tqdm==4.48.2
transformers==3.0.2
Werkzeug==1.0.1